Refactor ColorSwatchContainer to not be connected to the store
The parent component should pass down a handler for when a user clicking
the expand color icon. The actionType is now mapped to PaletteContainer.
