+++ src/components/core/ColorSwatch/ColorSwatch.sass
